lbavm012: fix transcoding problem.
tl;dr - decoding and encoding weren't symmetric for uimm20.

All operands except uimm20 were decoded into a normalized form at the
bottom of the uint32 that contained them. Having uimm20 as an exception
to this rule was convenient, but it caused problems when transcoding as
the encoding assumed normalized input.